Refactor `check_table()`
`check_table()` was an internal function that ensured that a dataframe was converted to a data table by reference, i.e. via `data.table::setDT()`.

This functionality was simplified to calling `data.table::setDT()` whenever `check_table()` was used. This update carries a slight speed improvement. But the real benefit is code simplification.